SAM-1347: Importing assessment with (large) attachments as ZIP file fails with error message
git-svn-id: https://source.sakaiproject.org/svn/sam/trunk@97673 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ samigo/samigo-app/src/java/org/sakaiproject/tool/assessment/ui/bean/delivery/DeliveryBean.java
@@ -249,7 +249,7 @@ public class DeliveryBean
   private Long assessmentGradingId;
   
   private boolean fromTableOfContents;
-  private float fileUploadSizeMax;
+  private int fileUploadSizeMax;
   private boolean studentRichText;
   
   // If set to true, delivery of the assessment is blocked.
@@ -3230,12 +3230,12 @@ public class DeliveryBean
 	    return s;
 	  }
 	  
-	  public void setFileUploadSizeMax(float fileUploadSizeMax)
+	  public void setFileUploadSizeMax(int fileUploadSizeMax)
 	  {
 	    this.fileUploadSizeMax = fileUploadSizeMax;
 	  }
 	  
-	  public float getFileUploadSizeMax()
+	  public int getFileUploadSizeMax()
 	  {
 	      return fileUploadSizeMax;
 	  }
